MOODLE_264

Conflicts:
	auth/email/auth.php
	badges/renderer.php
	calendar/renderer.php
	mod/scorm/locallib.php
	theme/afterburner/style/afterburner_dock.css
	theme/anomaly/style/dock.css
	theme/base/style/filemanager.css
	theme/bootstrapbase/javascript/dock.js
	theme/bootstrapbase/less/moodle/dock.less
	theme/bootstrapbase/style/moodle.css
	theme/clean/config.php
	theme/overlay/style/core.css
	theme/splash/style/pagelayout.css
	user/edit.php

+++ lib/cronlib.php
@@ -103,7 +103,7 @@ function cron_run() {
             $rs = $DB->get_recordset_sql ("SELECT *
                                              FROM {user}
                                             WHERE confirmed = 0 AND firstaccess > 0
-                                                  AND firstaccess < ?", array($cuttime));
+                                                  AND firstaccess < ? AND deleted = 0", array($cuttime));
             foreach ($rs as $user) {
                 delete_user($user); // we MUST delete user properly first
                 $DB->delete_records('user', array('id'=>$user->id)); // this is a bloody hack, but it might work
@@ -856,8 +856,11 @@ function cron_delete_from_temp() {
             // Check if file or directory is older than the given time.
             if ($iter->getMTime() < $time) {
                 if ($iter->isDir() && !$iter->isDot()) {
-                    if (@rmdir($node) === false) {
-                        mtrace("Failed removing directory '$node'.");
+                    // Don't attempt to delete the directory if it isn't empty.
+                    if (!glob($node. DIRECTORY_SEPARATOR . '*')) {
+                        if (@rmdir($node) === false) {
+                            mtrace("Failed removing directory '$node'.");
+                        }
                     }
                 }
                 if ($iter->isFile()) {
